Moving functions to utils
